pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.52.0",
    "langgraph>=1.0.4",
    "litellm>=1.80.7",
    "openai-agents>=0.6.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
